requirements.txt
PyQt6==6.8.1
PyQt6-Qt6==6.8.2
opencv-contrib-python==4.11.0
numpy==2.2.4
pyserial==3.5
